fixes of: * date range scenario, * one day scenario, * tag names, *

+++ src/main/scala/thesebas/spark/UrlRow.scala
@@ -10,9 +10,9 @@ case class UrlRow(
                    Url: String,
                    Date: Date,
                    tags: Set[String],
-                   var revs: Map[String, Float],
-                   var rev: Float,
-                   pic: Float,
-                   pif: Float,
+                   var revs: Map[String, Double],
+                   var rev: Double,
+                   pic: Double,
+                   pif: Double,
                    pi: Int
                  )
